Add kubernetes secrets sync script
This script takes a service, iterates through the paasta secrets on disk
and "syncs" them into kubernetes secrets. These are then native k8s
secrets that can be referenced as env vars/create tmpfs volumes.

I'm also saving the secret signature in a ConfigMap. This is so whatever
is watching for changes in secrets doesn't need to watch the secret
itself (and hence have access) to the secret to know if it needs to
bounce the service in question.